Fixing bug in SDF parsing

+++ iviz_urdf/Sdf/Box.cs
@@ -4,13 +4,13 @@ namespace Iviz.Sdf
 {
     public sealed class Box
     {
-        public Vector3f Scale { get; } = Vector3f.One;
+        public Vector3d Scale { get; } = Vector3d.One;
 
         internal Box(XmlNode node)
         {
             if (node.ChildNodes.Count != 0 && node.ChildNodes[0].Name == "scale")
             {
-                Scale = new Vector3f(node.ChildNodes[0]);
+                Scale = new Vector3d(node.ChildNodes[0]);
             }
         }
     }
